DC-15 Added
The DC-15 has been added. This is the first gun that has its own class, which essentially means that it has new features that we can use to create unique weapons. In a later date, we will change the different weapons to make them more unique, so there are not just a rehash of the E-11. For some reason the effects are not working, hopefully Linken will fix it.