DH-17 has a new class.
DH-17 has been updated to be a much more powerful blast but is less accurate then the E-11 and it uses 10 ammunition as well. You must update the weapons.dat for the damage to work if you are the player trying to kill an enemy. Look at the weapons.dat to see how I did it.